requirements.txt
flask
pypdf
dashscope
python-dotenv
